requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
qdrant-client>=1.7.0
cohere>=4.0.0
pydantic>=2.0.0